mini-apps/mae/requirements.txt
# Dependências do App Mãe
requests>=2.28.0
python-socketio>=5.8.0
Pillow>=9.0.0
opencv-python>=4.6.0
numpy>=1.21.0
pyautogui>=0.9.50
pytesseract>=0.3.10
